extracted product mobile and desktop version
